Rename name.py to name_parser.py
