+++ pb_plugins/dcsdkgen/name_parser.py
